namazu/requirements.txt
obspy
numpy
logging
matplotlib
pygame
pyserial
pyyaml
